First cut of Arms, rendering item IDs to every item file

+++ src/main/kotlin/sim/rotation/criteria/BuffDurationGte.kt
@@ -11,8 +11,8 @@ class BuffDurationGte(data: Map<String, String?>) : Criterion(Type.BUFF_DURATION
         null
     }
 
-    val seconds: Int? = try {
-        (data["seconds"] as String).toInt().coerceAtLeast(0)
+    val seconds: Double? = try {
+        (data["seconds"] as String).toDouble().coerceAtLeast(0.0)
     } catch (e: NullPointerException) {
         logger.warn { "Field 'seconds' is required for criterion $type" }
         null
