+++ src/main/kotlin/sim/rotation/criteria/DebuffDurationLte.kt
@@ -11,8 +11,8 @@ class DebuffDurationLte(data: Map<String, String?>) : Criterion(Type.DEBUFF_DURA
         null
     }
 
-    val seconds: Int? = try {
-        (data["seconds"] as String).toInt().coerceAtLeast(0)
+    val seconds: Double? = try {
+        (data["seconds"] as String).toDouble().coerceAtLeast(0.0)
     } catch (e: NullPointerException) {
         logger.warn { "Field 'seconds' is required for criterion $type" }
         null
