+++ src/main/kotlin/sim/rotation/criteria/FightTimeRemainingModuloLte.kt
@@ -4,8 +4,8 @@ import sim.SimIteration
 import sim.rotation.Criterion
 
 class FightTimeRemainingModuloLte(data: Map<String, String?>) : Criterion(Type.FIGHT_TIME_REMAINING_MODULO_LTE, data) {
-    val modulusSeconds: Int? = try {
-        (data["modulusSeconds"] as String).toInt().coerceAtLeast(0)
+    val modulusSeconds: Double? = try {
+        (data["modulusSeconds"] as String).toDouble().coerceAtLeast(0.0)
     } catch (e: NullPointerException) {
         logger.warn { "Field 'modulusSeconds' is required for criterion $type" }
         null
@@ -14,8 +14,8 @@ class FightTimeRemainingModuloLte(data: Map<String, String?>) : Criterion(Type.F
         null
     }
 
-    val seconds: Int? = try {
-        (data["seconds"] as String).toInt().coerceAtLeast(0)
+    val seconds: Double? = try {
+        (data["seconds"] as String).toDouble().coerceAtLeast(0.0)
     } catch (e: NullPointerException) {
         logger.warn { "Field 'seconds' is required for criterion $type" }
         null
